+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - An issue where the menus of window buttons became misaligned on Windows
 - Updated non-breaking dependencies
 - Updated Electron to 41.1.1
+- Implemented bridge.caspar.transform
 
 ## 1.0.0-beta.10
 ### Added